changed angular project name
