+++ PokeManiaWebApp/.editorconfig
