+++ PokeManiaWebApp/.gitignore
