+++ PokeManiaWebApp/README.md
@@ -1,4 +1,4 @@
-# PokeMania
+# PokeManiaWebApp
 
 This project was generated with [Angular CLI](https://github.com/angular/angular-cli) version 8.3.17.
