+++ PokeManiaWebApp/e2e/src/app.e2e-spec.ts
@@ -10,7 +10,7 @@ describe('workspace-project App', () => {
 
   it('should display welcome message', () => {
     page.navigateTo();
-    expect(page.getTitleText()).toEqual('PokeMania app is running!');
+    expect(page.getTitleText()).toEqual('PokeManiaWebApp app is running!');
   });
 
   afterEach(async () => {
