+++ PokeManiaWebApp/e2e/tsconfig.json
