+++ PokeManiaWebApp/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "poke-mania",
+  "name": "poke-mania-web-app",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
