+++ PokeManiaWebApp/src/app/app.component.spec.ts
@@ -16,16 +16,16 @@ describe('AppComponent', () => {
     expect(app).toBeTruthy();
   });
 
-  it(`should have as title 'PokeMania'`, () => {
+  it(`should have as title 'PokeManiaWebApp'`, () => {
     const fixture = TestBed.createComponent(AppComponent);
     const app = fixture.debugElement.componentInstance;
-    expect(app.title).toEqual('PokeMania');
+    expect(app.title).toEqual('PokeManiaWebApp');
   });
 
   it('should render title', () => {
     const fixture = TestBed.createComponent(AppComponent);
     fixture.detectChanges();
     const compiled = fixture.debugElement.nativeElement;
-    expect(compiled.querySelector('.content span').textContent).toContain('PokeMania app is running!');
+    expect(compiled.querySelector('.content span').textContent).toContain('PokeManiaWebApp app is running!');
   });
 });
