+++ PokeManiaWebApp/src/polyfills.ts
