+++ PokeManiaWebApp/tsconfig.json
@@ -10,7 +10,7 @@
     "module": "esnext",
     "moduleResolution": "node",
     "importHelpers": true,
-    "target": "esnext",
+    "target": "es2015",
     "typeRoots": [
       "node_modules/@types"
     ],
