+++ PokeManiaWebApp/tslint.json
